fix explicit conversion error

+++ Chapter03/CastingConverting/Program.cs
@@ -4,6 +4,6 @@ double b = a; // An int can be safely cast into a double.
 WriteLine($"a is {a}, b is {b}");
 
 double c = 9.8;
-int d = c; // Compiler gives an error if you do not explicitly cast.
-WriteLine($"c is {c}, d is {d}");
+int d = (int)c; // Compiler gives an error if you do not explicitly cast.
+WriteLine($"c is {c}, d is {d}"); // d loses the .8 part.
 #endregion
